pyproject.toml
[project]
name = "json-format"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "groq>=1.0.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.2.1",
    "langchain-groq",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
]
